Replace ConnectModule with WalletModule
