--- magic/core/src/main/java/link/magic/android/modules/connect/requestConfiguration/RequestUserInfoConfiguration.kt
@@ -1,6 +1,0 @@
-package link.magic.android.modules.connect.requestConfiguration
-
-import androidx.annotation.Keep
-
-@Keep
-data class RequestUserInfoConfiguration(var isResponseRequired: Boolean = false)
