+++ magic/core/src/main/java/link/magic/android/modules/wallet/requestConfiguration/RequestUserInfoWithUIConfiguration.kt
@@ -1,0 +1,16 @@
+package link.magic.android.modules.wallet.requestConfiguration
+
+import androidx.annotation.Keep
+import org.web3j.protocol.core.Response
+
+@Keep
+data class RequestUserInfoWithUIConfiguration(var scope: WalletUserInfoScope)
+
+@Keep
+data class WalletUserInfoScope(var email: WalletUserInfoEmailOptions)
+
+@Keep
+enum class WalletUserInfoEmailOptions {
+    required, optional
+}
+
