+++ magic/core/src/main/java/link/magic/android/modules/wallet/response/ConnectWithUIResponse.kt
@@ -1,0 +1,7 @@
+package link.magic.android.modules.wallet.response
+import org.web3j.protocol.core.Response
+
+import androidx.annotation.Keep
+
+@Keep
+class ConnectWithUIResponse: Response<Array<String>>()
